requirements_prod.txt
Django==5.0.6
Pillow==10.4.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
django-environ==0.11.2
python-dotenv==1.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
